backend/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.8
python-snap7==1.3
Werkzeug==2.3.6
setuptools>=65.5.1
gunicorn==20.1.0
Flask-SocketIO==5.3.6
eventlet==0.33.3
urllib3==2.0.7
gevent==23.9.1
gevent-websocket==0.10.1
APScheduler
requests==2.31.0
xhtml2pdf==0.2.16
reportlab>=4.0.4,<5
pytz==2023.3
asteval==0.9.31